Fixed some OAuth bugs, reimplemented github issues, refactored templates a bit
- OAuth sign in / login refresh now returns user to their last
page
- login_required decorator used is now the flask_login one,
meaning it can be disabled for testing purposes
- Refreshed logins now work properly
- Refactored some html into a 'collapsible_section' snippet. Things
are a bit more readable and DRY, but page loads are slower. May be
due to a template caching issue. Will investigate later.
- Added some support functions for github issue searching and
creation, and added a new section to the timepoint page to
display them. Styling needs massive work, and may want to add
issue body later, but it's functional!
- Reimplemented the 'create_issue' page